requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-imagekit==4.1.0
Pillow==10.1.0
drf-yasg==1.21.7
django-cors-headers==4.3.1